% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwthcolors.R
\name{rwth_cols}
\alias{rwth_cols}
\title{Function to extract rwth colors as hex codes}
\usage{
rwth_cols(...)
}
\arguments{
\item{...}{Character names of rwth_colors}
}
\description{
Function to extract rwth colors as hex codes
}
