% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwthcolors.R
\name{rwth_pal}
\alias{rwth_pal}
\title{Return function to interpolate a rwth color palette}
\usage{
rwth_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in rwth_palette}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a rwth color palette
}
